Fix audit issues: stack size, HA poll sleep, WiFi reconnect, minor cleanups
- Increase UI task stack from 2048 to 4096 bytes to prevent overflow
  from FastEPD display operations and font rendering
- Replace HA task 10ms busy-poll with notification-based wake using
  SLEEP_WAKE_INTERVAL_MS timeout — biggest remaining power fix, allows
  CPU to actually sleep between WebSocket service intervals
- Add explicit WiFi.setAutoReconnect(true) instead of relying on
  implicit default behavior
- Remove unnecessary int16_t casts in screen.cpp Rect initialization
- Fix unsigned comparison in accumulate_damage (== 0 vs <= 0)

+++ src/screen.cpp
@@ -9,10 +9,10 @@ void screen_add_slider(SliderConfig config, Screen* screen) {
     }
 
     Rect rect{
-        .x = (int16_t)config.pos_x,
-        .y = (int16_t)config.pos_y,
-        .w = (int16_t)config.width,
-        .h = (int16_t)config.height,
+        .x = config.pos_x,
+        .y = config.pos_y,
+        .w = config.width,
+        .h = config.height,
     };
 
     Slider* widget = new (std::nothrow) Slider(config.label, config.icon_on, config.icon_off, rect);
@@ -31,10 +31,10 @@ void screen_add_button(ButtonConfig config, Screen* screen) {
     }
 
     Rect rect{
-        .x = (int16_t)config.pos_x,
-        .y = (int16_t)config.pos_y,
-        .w = 0, // ignored
-        .h = 0, // ignored
+        .x = config.pos_x,
+        .y = config.pos_y,
+        .w = 0,
+        .h = 0,
     };
 
     OnOffButton* widget = new (std::nothrow) OnOffButton(config.label, config.icon_on, config.icon_off, rect);
